arm7: merge SND_channel into SND_exChannel
