+++ arm7/lib/src/SND_command.c
@@ -10,7 +10,6 @@
 #include "SND.h"
 #include "SND_alarm.h"
 #include "SND_capture.h"
-#include "SND_channel.h"
 #include "SND_exChannel.h"
 #include "SND_main.h"
 #include "SND_seq.h"
